Fix "genstrings" issue
When exporting, xcodebuild tool was giving an error that the string in
the macro wasnt a literal string? But they were string literals. This
caused the xliff to not have the strings.  On a
whim, tried adding a second arg (to be closer to NSlocalizedString) and
it got rid of the error and the xliff had the strings in it again